requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.7
botocore==1.35.7
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
idna==3.8
importlib_resources==6.4.4
imutils==0.5.4
jmespath==1.0.1
kiwisolver==1.4.6
matplotlib==3.9.2
multidict==6.0.5
numpy==2.0.2
openai==0.28.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.6
yarl==1.9.4
zipp==3.20.1
